+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This can also be enabled programmatically with `warnings.simplefilter('default',
 - `arc()` not longer renders artefacts at intersection point, thanks to @Jmillan-Dev; [#488](https://github.com/PyFPDF/fpdf2/issues/488)
 ### Added
 - workaround by @semaeostomea to support arabic and right-to-left scripts: [link to documentation](https://pyfpdf.github.io/fpdf2/Unicode.html#arabic-script-workaround)
+- documentation on shapes styling [link to documentation](https://pyfpdf.github.io/fpdf2/Shapes.html#path-styling)
 
 ## [2.5.6] - 2022-08-16
 ### Added